renamed files per spec & fixed loggin error
